Support TextInput::onKeyPress
closes #413

onKeyPress is not supported natively, however it can be done by combining `change`, `textSelectionChange` and `paste/cut` events, as it is done on other platforms. 

Currently the behaviour differs from Android in 1 scenario - when `backspace` is pressed in an empty TextInput, the key press will not be recorded as no event is fired in that situation in ArkUI component.

See merge request rnoh/react-native-harmony!551

Signed-off-by: huangyouhua <1356795413@qq.com>