Implemented TextInput::allowFontScaling + test - but as we can only change...
Implemented TextInput::allowFontScaling + test - but as we can only change displaySize (with no way of changing textSize) we cannot reliably check that.
Closes #562

While the code will set `fp` for when `allowFontScaling=true` and `vp` when `allowFontScaling=false`, we have no way of seeing the results, because on devices we can only change displaySize, and there is no way of changing fontSize. So both `fp` and `vp` behaves the same.

See merge request rnoh/react-native-harmony!528

Signed-off-by: huangyouhua <1356795413@qq.com>