feature: codegen metro sever ip
This PR codegens METRO_SERVER_IP. `localhost` doesn't work (at least for me), because Metro sever is not hosted by the phone. On iOS, `react-native-xcode.sh` is responsible for providing this IP. OpenHarmony however, can be developed on Windows and on Macs, so we need to have a platform agnostic solution. I decided to use Node.js since it's a dependency of React Native and OpenHarmony. I also wanted to share the codegen code across projects, so I moved it to the CLI package. The command responsible for generating this file, is triggered from CMakeFile.

Related https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/369

See merge request rnoh/react-native-harmony!676

Signed-off-by: huangyouhua <1356795413@qq.com>