Remove TextInput::onKeyPress support to fix TextInput::onChange
API 11 changes introduced a bug where ArkUI `onTextSelectionChange` callback would not fire when caret position changed but nothing was actually selected. This causes the caret position to be not determinable, and as a result `TextInput::onKeyPress` cannot be implemented, as it relies on the caret position to determine the most recently pressed key. 

Because changing the value of the TextInput also changes the caret position, the `TextInput::onChange` was invoked from the `onTextSelectionChange` callback to provide correct callback order (first `onKeyPress` then `onChange`).

Since `onChange` is an indispensable feature the code for `onKeyPress` is removed to allow that functionality to work. These changes can be reverted in the future when either `onTextSelectionChange` is fixed or some other method of determining caret position is introduced.

closes #735

See merge request rnoh/react-native-harmony!677

Signed-off-by: huangyouhua <1356795413@qq.com>