Fix scrollToEnd not triggering scrolling when invoked after onLayout
`scrollToEnd` command was not always triggering the scrolling when the window dimension change caused the layout to change. Replacing the native method `scrollEdge` with `scrollTo` fixes that problem. The reverse of this change was done previously to fix a bug, that when checked now is no longer appearing so this should be fine.

See merge request rnoh/react-native-harmony!795

Signed-off-by: huangyouhua <1356795413@qq.com>