Support TextInput::onKeyPress (without backspace)
After some fixes delivered with API 11 `TextInput::onKeyPress` can now be partially supported. At the moment pressing backspace will not trigger the `keyPress` event as it is indistinguishable from cutting the text using context menu. This is due to the `onCut` callback arriving after `onTextSelectionChange` which is used to determine which key was pressed.

related #736

See merge request rnoh/react-native-harmony!822

Signed-off-by: huangyouhua <1356795413@qq.com>