ScrollView::keyboardDismissMode and ScrollView:: keyboardShouldPersistTaps
Moved the test for `ScrollView::keyboardShouldPersistTaps` to the examples. All of our tests are nested inside a ScrollView. In order to `keyboardShouldPersistTaps` work in nested `<ScrollView />` we need to set the same value on parent `<ScrollView />`.

Added patch to support `keyboardDismissMode` for OpenHarmony in React Native.

Closes: https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/311 and https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/310

See merge request rnoh/react-native-harmony!834

Signed-off-by: huangyouhua <1356795413@qq.com>