refactor: Use refcounting for component managers
Previously, each Ark component registered its own `ComponentManager`.
`ComponentManagerRegistry` kept track of all `Manager`s registered for a tag, and returned the most recently created one.

With this PR, `ComponentManager`s are reused between Ark components with the same tag.
This is accomplished by keeping a reference counter for registered `ComponentManager`s.
When a component first appears, it increments the refcounter and retrieves the registered `Manager` if one exists, and registers a new one otherwise.
When a component disappears, it decrements the refcounter and destroys the manager if it reached 0.
This removes the need for components to remember their `ComponentManager`, which was previously necessary to unregister them -- now they can use the tag for unregistering.

See merge request rnoh/react-native-harmony!842

Signed-off-by: huangyouhua <1356795413@qq.com>