Force Scroll component to start at the content offset from state if provided
closes #872 

Previously the scroll position would not be maintained when user navigated to a different page and back using `react-navigation` because the ArkUI component would be destroyed and created again. Since this information is kept in `ScrollViewState` then we can manually scroll to remembered offset from state to mimick maintaing the offset.

This MR also fixes a crash which was caused by trying to retrieve `scroller.contentOffset` when the component was not fully mounted on the native side.

See merge request rnoh/react-native-harmony!867

Signed-off-by: huangyouhua <1356795413@qq.com>