test: onResponderReject
OnResponerReject requires both components to become responders, but in the current test case, only the child component is a responder, and the parent component will not become a responder. So onResponderReject will never be triggered.
Here, I set the parent component as the responder of move, so that the trigger condition of onResponderReject is met, and onResponderReject can be verified.

See merge request rnoh/react-native-harmony!885

Signed-off-by: huangyouhua <1356795413@qq.com>