fix : textinput in multiline mode hidde border and set default selectioncolor.
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
When the selection color is not set, the selection color is set to 0x7fffffff. When CppComponentInstance::onPropsChanged(props) is executed, the mutiline\_ is not set to true, resulting in the textinput not displaying the border in multiline mode
## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->
add new function resetSelectedBackgroundColor
## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. Run "TextinputTest::render textinput with default selection color" and "TextinputTest::render multiline text input with common border"


## Checklist
- [ ] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1025

Signed-off-by: huangyouhua <1356795413@qq.com>