fix: caret flash
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
When the TextInput component uses onChange to change the Value to change the content, if the modification is too fast, the caret will appear in the wrong position, especially when it is deleted, it is easier to reproduce.
This will prevent all text from being deleted.

## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. 


## Checklist
- [ ] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1036

Signed-off-by: huangyouhua <1356795413@qq.com>