fix: handle command race with MAIN when retrieving componentInstance
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
Handle component commands fully on MAIN thread

## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->
Currently, if a view is CREATEd and immediately a command is dispatched to it, it's possible command handling will fail to find that view in `componentInstanceRegistry`.

This moves the retrieval of `componentInstance` to the MAIN thread to ensure the view can be found (thanks to the sequential running of scheduled tasks).

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. Go to VirtualizedList tests
3. Verify `scrollTo`/`initialIndex`/other command-related functionalities still work.


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
-->
Closes #976

See merge request rnoh/react-native-harmony!1046

Signed-off-by: huangyouhua <1356795413@qq.com>