fix: crash while deleting tag
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
After MR#1003, the mix components are encountering an issue that the componentInstance could possibly not be created and inserted in our map, and while trying to delete it the APP will crash.
## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->
`react-native-harmony/harmony/cpp/RNOH/ComponentInstanceRegistry.h`
## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. 


## Checklist
- [ ] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1087