AlertManagerTurboModule: ensure that onAction is called only once
## Summary

This PR fixes #1046. The reason behind this bug is the behavior of `onAction` dialog button callback in ArkUI - if clicked multiple times in short period of time, the callback is executed more than once, even though the button has started the hide animation. This PR introduces a small workaround which sets the `onAction` callback to undefined, to ensure that it's not called more than once.  

## Changes
- ensure calling the onAction callback only once in AlertManagerTurboModule

## Test Plan
1. Create a custom component which shows an alert.
2. Try to click the confirm button multiple times in a short period of time - the app should not crash.


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes: #1046.

See merge request rnoh/react-native-harmony!1176

Signed-off-by: huangyouhua <1356795413@qq.com>