Fix ScrollView bugs
## Summary
This fixes two `ScrollView` connected bugs. Firstly, the nested scrolling mode was unnecessarily updated on every contentSize and containerSize change - the mode is updated on each onScroll either way. This caused problems with the initial settings as some of the info needed for updating the mode was not available at component initialisation. Secondly, once `snapToOffsets` or `snapToInterval` was set it could not be unset as the case when both of these were not given was not handled.


## Changes
- fix `snapToOffsets` and `snapToInterval` still being applied after unsetting them
- fix nested ScrollViews triggering parent scrolling when scrolled beyond its content right after initialisation


## Test Plan
1. Open tester app.
2. Go to `NestedScrollingExample`, scroll the parent to see the child then scroll the child with a strong fling and observe it stop on its boundary without triggering scrolling in the parent


## Checklist
- [X] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1177

Signed-off-by: huangyouhua <1356795413@qq.com>