fix: issue with JS callbacks being dropped when creating a new RN Instance
## Summary
JS callbacks passed to ArkTS were retained as `LongLivedObjects`. Due to a bug in RN, when a new instance is created, the `LongLivedObjectsCollection` is cleared, causing the callback to be dropped before being called. This PR retains the callback in the object passed to ArkTS.

## Changes
- fixed JS callbacks passed to ArkTS being cleared prematurely

## Issue Links
Closes #1058

See merge request rnoh/react-native-harmony!1211

Signed-off-by: huangyouhua <1356795413@qq.com>