StatusBar: change default background color to transparent
## Summary
This PR changes the default background color of `StatusBar` from semi-transparent to transparent. This change is mostly motivated by the default behavior of `statusBarColor` in OH `SystemBarProperties` - when unset, the background is actually transparent, it's not set to semi-transparent as stated in the [docs](https://docs.openharmony.cn/pages/v4.1/en/application-dev/reference/apis-arkui/js-apis-window.md). The StatusBar color is also set to transparent on iOS and, from what I tested, on Android (however the default background color on Android is set by the system, so there might be some other behaviors depending on system platform). 

## Changes
- changed the default background color of StatusBar

## Test Plan
1. Tested what's the default behavior of statusBarColor in native OH app. 
2. Tested if the background color of StatusBar is transparent by default. 

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Related: #1070

See merge request rnoh/react-native-harmony!1226