@sosen/1072 implement and test missing xmlhttprequest features
## Summary
This MR adds the upload progress event, aswell as fixing the timeout property - to align better with XMLHttpRequests timeout. (The native timeouts were for specific parts of the request, while the XMLHttpRequest timeout is for the whole request). Also, the onTimeout event was fixed.

## Changes
- added upload progress event handling
- fixed timeout event


## Test Plan
1. Open tester app.
2. Check Networking tests


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Related: #946 
Closes: #1072

See merge request rnoh/react-native-harmony!1234