fix: load bundle crash with platform error
<!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
This PR fixes the crashing that when the app is trying to load bundles, it would crash when loading fails.

## Changes
<!-- List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

For breaking changes, use "[breaking]" prefix, for example:
- [breaking] removed something
-->
Add initialization for napi_value, so that even the loading fails the app would not crash. 

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. 


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1261

Signed-off-by: huangyouhua <1356795413@qq.com>