fix: SafeAreaView not taking effect when it should
## Summary
In my understanding, SafeAreaView is a view with padding added. Sometimes, SafeAreaView does not work because obtaining the width of the borders around the screen sometimes fails. I found that sometimes the getLastWindow API returns an exception for BusinessError (I cannot find the specific reason), so I added a check. When it throws a BusinessError, I will call the getLastWindow API again. Such changes may not be perfect enough, but currently they can solve the problem of SafeAreaView not taking effect.

## Changes
- fixed SafeAreaView not working in certain situations

## Test Plan
No testing plan

## Checklist
- [x] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1241

Signed-off-by: huangyouhua <1356795413@qq.com>