chore: update styleguide + refactor: improve naming consistency
## Summary
The current style guide rules are inconsistent. They recommend using "arkJs" or "arkTs" instead of "arkJS" and "arkTS." To maintain consistency, we would need to apply a similar format to "arkUi," but no one writes "UI" this way. On the other hand, we use "Id" but not "ID." However, "Id" is an abbreviation, while "UI" is an acronym.

Are "JS" and "TS" acronyms or abbreviations? I am not entirely sure, but I think they should follow the rules for acronyms. We already adhere to this approach when naming files and classes, but we don't apply it in camel case. Also, ArkTS wiki page uses this casing.

This PR renames "arkJs" to "arkJS" and "arkTs" to "arkTS" to ensure consistency.

## Changes
- deprecated `ComponentInstance::Depenedencies::arkTsMessageHub` in favor of `arkTSMessageHub`

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1124

See merge request rnoh/react-native-harmony!1306

Signed-off-by: huangyouhua <1356795413@qq.com>