TextInput Props Update
<!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
This PR synchronizes all the new properties and updates we made on TextInput from May on GitHub. It is a draft that commits all changes we have made to the code. more info for changes and texting cases will be updated accordingly.

## Changes
<!-- List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

For breaking changes, use "[breaking]" prefix, for example:
- [breaking] removed something
-->
[breaking] move convertTextAreaInputType, convertInputType, convertEnterKeyType, convertContentType to corresponding ArkUI node. 

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. Run TextInput tests and observes them work


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1414

Signed-off-by: huangyouhua <1356795413@qq.com>