fix: use folly::Function instead of std::function in TaskExecutor
## Summary
Previously, `Task`s were an alias to `std::function` which needs to be copyable. We often want to move resources into the tasks we schedule, which is made more error-prone when the `Task` can be copied. This should help ensure the resources held by `Task`s are handled correctly.

## Changes
- changed `TaskExecutor::Task` to only require the passed object to be moveable, not copyable

See merge request rnoh/react-native-harmony!1463

Signed-off-by: huangyouhua <1356795413@qq.com>