fix: cleanup task runners on a separate thread
## Summary
When destroying the `TaskExecutor` we need to schedule a special "cleanup" task on each task runner in that executor and wait for its completion.
Sometimes, when the `TaskExecutor` is freed on the MAIN thread, the WorkerThread is blocked on an ArkTS API call which internally blocks and waits for the MAIN thread, resulting in a deadlock.
This PR awaits the completion of the "cleanup" tasks on the CLEANUP thread, resolving the deadlock.

## Changes
- fixed possible deadlock when destroying a RN Instance

##
Related to #1217

See merge request rnoh/react-native-harmony!1485

Signed-off-by: huangyouhua <1356795413@qq.com>