optimization: make Paragraph a leaf shadow node
## Summary
In RN, the `ParagraphShadowNode` has the shadow node trait `FormsStackingContext`, which means that the corresponding `ShadowView` can have children.

On Android, that flag is disabled, since the Android `TextView` cannot have children. Since ArkUI's `Text` node is the same in this matter, it makes sense to unset that trait in RNOH as well.

This also has the advantage of flattening the ShadowView tree, which seems to improve the `Differentiator` performance in real-world scenarios.

## Changes
- optimization: flatten the components nested within `Text` in the native view hierarchy

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. Verify no regressions are present in the `Text` test suite


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes #1229

See merge request rnoh/react-native-harmony!1490

Signed-off-by: huangyouhua <1356795413@qq.com>