feat: interpolate strings in animations + refactor: ValueAnimatedNode
## Summary
This PR adds support for native string interpolation, which can be used by svg packages in order to animate the `d` property. This PR also clarifies and generalizes the concept of value in order to support strings and numbers, and introduces the concept of output (output = value + offset (offset is applicable if value is double)).

## Changes
- added string interpolation for native animations

## Test Plan
Check "Animated::interpolate::interpolate numbers detected in a string" test.

## Checklist
- [x] Project compiles and the tester application works.
- [x] Fix "Error in animation update: There were 17 active nodes, but only 15 were updated"

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1113

See merge request rnoh/react-native-harmony!1500

Signed-off-by: huangyouhua <1356795413@qq.com>