OHReactMarkerListener: fix fabric markers and logMarker method
## Summary
This PR fixes logMarker method and how fabric markers are handled in OHReactMarkerListener - previously timestamp was placed in a place of a tag - because of that, the message created in logMarkerStart was different than in logMarkerEnd, causing the start marker to never find the corresponding end marker. 

## Changes
- fixed never ending React Markers in a trace files

## Test Plan
After applying this MR, check if fabric markers have an end and if markers logged with logMarker are visible in a trace file (e.g. CONTENT_APPEARED)

## Checklist
- [x] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1542

Signed-off-by: huangyouhua <1356795413@qq.com>