feat: adjust RNSurface size to its content
## Changes
- added `RNSurface::constraints` and ability of adjusting dimensions of `RNSurface` to its content

## Test Plan
Special testing environment that changes RNSurface dimension on ArkUI side, and changes RootComponent dimensions on RN side. That environment required patching RN — flex: 1 needs to be removed from AppContainer.

### Custom Constraints

`RNSurface::constraints` property allow the surface to expand beyond the area specified using .width() and .height() properties. You can use .width() and .height() to specify initial size, and RNSurface::constraints to specify max (and min) size. Notice that changing height doesn't impact the size of the surface because provided constraints are used.

![custom_constraints](/uploads/79656e20935a6409ccc825cd57cd6a10/custom_constraints.mp4)

### No Custom Constraints

If `RNSurface::constraints` is not specified, then constraints provided by ArkUI are used (with the exception of minWidth, because I couldn't expand the surface in horizontal direction on RN side). If the component inside RN is larger than RNSurface, the content of RNSurface will overflow the container. In this video, changing the .height() property impacts the content in RNSurface (if that content has dynamic dimensions).

![no_custom_constraints](/uploads/02978fc6844419c4fcf3d3d246ee6a36/no_custom_constraints.mp4)


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1226

See merge request rnoh/react-native-harmony!1563

Signed-off-by: huangyouhua <1356795413@qq.com>