fix: deadlock when calling surface methods
## Summary
Surface operations can cause mutations, which in turn can call the mutation listener callback passed to `MountingManagerArkTS`.
Because the bridging functions from RNOHAppNapiBridge locked the `RNInstance`s mutex for their entire duration, and the mutation listener locked it to check if the Instance was still alive, we had the same thread trying to lock the mutex twice in that case, leading to the application hanging.

## Changes
- fixed deadlock when calling `SurfaceHandle` methods

## Test Plan
1. Open multi_surface app.
2. Go to any example (except "standalone app") and press "Show Surface"
3. Verify the app doesn't hang.

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes #1277

See merge request rnoh/react-native-harmony!1577

Signed-off-by: huangyouhua <1356795413@qq.com>