fix: incorrect cursor placement when setting content through props
<!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->

This PR fixes the incorrect behavior of the cursor when the value of TextInput is set via state. Before those changes, the cursor moved to the end instead of staying in the same place with respect to the end of the word. 

## Changes
<!-- List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed 
- deprecated
- removed

For breaking changes, use "[breaking]" prefix, for example:
- [breaking] removed something
-->

- fixed incorrect cursor placement when setting content via state

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
-->
Closes #1265

See merge request rnoh/react-native-harmony!1565

Signed-off-by: huangyouhua <1356795413@qq.com>