fix: remove workaround and fix layouting of ArkTS components
## Summary
ArkTS component nodes aren't laid out correctly inside `CustomNode`. Currently, we use a hacky workaround of wrapping `FallbackComponentInstance`s in `RootComponentInstance`s (which use `StackNode`), which is a bad idea on many levels. This PR instead wraps the contents of `FallbackComponentInstance` in an extra `StackNode` and sets the layout on it.

## Changes
- removed workaround of wrapping Custom ArkTS components in `RootComponentInstance`s

## Test Plan
1. Open tester app.
2. Verify "Custom Native Component" tests still pass.


## Checklist
- [x] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1594

Signed-off-by: huangyouhua <1356795413@qq.com>