fix: call js timers once per frame
## Summary
iOS and Android trigger timers in frame callbacks. This PR brings us closer to other platforms.

This also improves performance when using `Animated` API with native driver disabled: this is becuase `requestAnimationFrame` is implemented as `setTimeout(..., 1)`, which in our implementation causes the updates to fire as often as each ms.

## Changes
- fix: trigger JS timers only once per frame
- optimization: trigger JS timers that finish at a similar time together, reducing the number of JSI calls

## Test Plan
1. Open tester app.
2. Go to Timer test suite.
3. Verify timers still work.


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes #1301

See merge request rnoh/react-native-harmony!1613

Signed-off-by: huangyouhua <1356795413@qq.com>