fix: TimingTurboModule crashes when destroyed
## Summary

Previously, we called `cancelWakeUp()` inside the `TimingTurboModule` destructor, which caused app crashes. The reason behind this is that the destructor was executed on the CLEANUP thread, while `cancelWakeUp()` checks if it runs on the JS thread with assertion. After changes, the destructor just calls `taskExecutor->cancelDelayedTask`.

## Changes

- fixed: incorrect thread assertion inside `TimingTurboModule` destructor

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links

Closes: #1324

See merge request rnoh/react-native-harmony!1641

Signed-off-by: huangyouhua <1356795413@qq.com>