maybe fix: interal redirects for scoped aliases in Metro config for Windows
## Summary
It looks like Metro config doesn't handle internal redirects properly because of differences in paths on Mac and Windows.

## Changes
- maybe fixed internal redirects for aliases with scoped package name on Windows

## Test Plan

Added a test case, but it was passing before the change was introduced. I am testing on Mac, but I suspect that test would fail on Windows.

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1344

See merge request rnoh/react-native-harmony!1670

Signed-off-by: huangyouhua <1356795413@qq.com>