fix:setState will trigger an onChange
## Summary
I have modified the method which of preventing setState from triggering onchange

## Changes
In the previous method, it used 'm_shouldIgnoreNextChangeEvent' to prevent setState from triggering onchange,but it will fail in a few case.It is difficult to implement this solution on the 'ComponentInstance' side, so I have implemented it on the Node side. When the 'onChange' event is triggered by 'setTextContent', I will prevent it from being passed to the 'ComponentInstance' side.

## Test Plan
1. Open tester app.
2. Run "TextInputTest::should log only once, after pressing space". 

## Checklist
- [ ] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1681

Signed-off-by: huangyouhua <1356795413@qq.com>