optimization: Remove whole ShadowView setting when preallocating components
## Summary
This PR removes the whole shadowView setting when preallocating components.

## Changes
- optimized rendering pipeline (10-15ms/1500 views)

## Test Plan
1. Open tester app.
2. Run Benchmark.

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
`SchedulerDelegate::schedulerDidRequestPreliminaryViewAllocation` is on the critical path of rendering. Constucting a `ShadowView` class is slower than a structure. For 1500 Views, Constucting a `ShadowView` takes 10~15ms more.

Removed the following 4 settings from preallocation
`
    componentInstance->setLayout(shadowView.layoutMetrics);
    componentInstance->setEventEmitter(shadowView.eventEmitter);
    componentInstance->setState(shadowView.state);
    componentInstance->setProps(shadowView.props);
`
The `layoutMetrics` is not yet measured when preallocating.
The `eventEmitter` and `state` doesn't take much time.
The `props` is more complicated.
Once we confirm that the `props` from `SchedulerDelegate::schedulerDidRequestPreliminaryViewAllocation` is the same as the one from `MountingManagerCAPI::handleMutation`, we should set
it during preallocation and skip the setting during handleMutation. (#1414)
@kasperski mentioned that it's okay to call `setProps` twice since !1357 can "skip" the second call. It works for the components we provide, but it is risky for third-party components.

See merge request rnoh/react-native-harmony!1714

Signed-off-by: huangyouhua <1356795413@qq.com>